BackOffice/pyproject.toml
[tool.poetry]
name = "backoffice"
version = "0.1.0"
description = ""
authors = ["Juan David Duque <jduque0289@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.17.2"
webdriver-manager = "^4.0.1"
requests = "^2.31.0"
bs4 = "^0.0.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
